Refactored and simplified proximity graph.
Instead of implementing our own graph methods,
I derived from an existing undirected graph class.
I'm not sure why it was not done that way in the
first place, so I suspect I missed something. We
should discuss this change in person.

That class obviously took a while to implement so
I'm sorry that I undoed this work.

I'm pretty sure the same can be done with Transmission
Graph.